refactor: dto를 presentation 계층에서 application 계층으로 이동
