+++ src/main/java/roomescape/service/ReservationService.java
@@ -19,10 +19,10 @@ import roomescape.exception.reservation.InvalidDateTimeReservationException;
 import roomescape.exception.reservation.NotFoundReservationException;
 import roomescape.exception.theme.NotFoundThemeException;
 import roomescape.exception.time.NotFoundTimeException;
-import roomescape.web.dto.request.reservation.ReservationRequest;
-import roomescape.web.dto.request.reservation.ReservationSearchCond;
-import roomescape.web.dto.response.reservation.ReservationResponse;
-import roomescape.web.dto.response.reservation.UserReservationResponse;
+import roomescape.service.dto.request.reservation.ReservationRequest;
+import roomescape.service.dto.request.reservation.ReservationSearchCond;
+import roomescape.service.dto.response.reservation.ReservationResponse;
+import roomescape.service.dto.response.reservation.UserReservationResponse;
 
 @Service
 @RequiredArgsConstructor
