+++ src/main/java/roomescape/service/dto/request/time/ReservationTimeRequest.java
@@ -1,17 +1,16 @@
-package roomescape.web.dto.response.time;
+package roomescape.service.dto.request.time;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonFormat.Shape;
+import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
 import roomescape.domain.ReservationTime;
 
-public record ReservationTimeResponse(Long id, LocalTime startAt) {
+public record ReservationTimeRequest(
+        @NotNull(message = "시간은 빈값을 허용하지 않습니다.") LocalTime startAt) {
 
-    public static ReservationTimeResponse from(ReservationTime time) {
-        return new ReservationTimeResponse(
-                time.getId(),
-                time.getStartAt()
-        );
+    public ReservationTime toReservationTime() {
+        return new ReservationTime(startAt);
     }
 
     @Override
