+++ src/main/java/roomescape/service/dto/response/member/MemberResponse.java
@@ -1,4 +1,4 @@
-package roomescape.web.dto.response.member;
+package roomescape.service.dto.response.member;
 
 public record MemberResponse(Long id, String name) {
