+++ src/main/java/roomescape/service/dto/response/reservation/ReservationResponse.java
@@ -1,10 +1,10 @@
-package roomescape.web.dto.response.reservation;
+package roomescape.service.dto.response.reservation;
 
 import java.time.LocalDate;
 import roomescape.domain.Reservation;
-import roomescape.web.dto.response.member.MemberResponse;
-import roomescape.web.dto.response.theme.ThemeResponse;
-import roomescape.web.dto.response.time.ReservationTimeResponse;
+import roomescape.service.dto.response.theme.ThemeResponse;
+import roomescape.service.dto.response.time.ReservationTimeResponse;
+import roomescape.service.dto.response.member.MemberResponse;
 
 public record ReservationResponse(Long id,
                                   LocalDate date,
