+++ src/main/java/roomescape/web/controller/AdminController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.service.ReservationService;
-import roomescape.web.dto.request.reservation.ReservationRequest;
-import roomescape.web.dto.response.reservation.ReservationResponse;
+import roomescape.service.dto.request.reservation.ReservationRequest;
+import roomescape.service.dto.response.reservation.ReservationResponse;
 
 @RestController
 @RequestMapping("/admin")
