+++ src/main/java/roomescape/web/controller/AuthController.java
@@ -14,8 +14,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.service.MemberService;
 import roomescape.service.security.JwtProvider;
-import roomescape.web.dto.request.member.LoginRequest;
-import roomescape.web.dto.response.member.MemberResponse;
+import roomescape.service.dto.request.member.LoginRequest;
+import roomescape.service.dto.response.member.MemberResponse;
 
 @RestController
 @RequiredArgsConstructor
