+++ src/main/java/roomescape/web/controller/ReservationController.java
@@ -14,12 +14,12 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.service.ReservationService;
-import roomescape.web.dto.request.member.MemberInfo;
-import roomescape.web.dto.request.reservation.ReservationRequest;
-import roomescape.web.dto.request.reservation.ReservationSearchCond;
-import roomescape.web.dto.request.reservation.UserReservationRequest;
-import roomescape.web.dto.response.reservation.ReservationResponse;
-import roomescape.web.dto.response.reservation.UserReservationResponse;
+import roomescape.service.dto.request.member.MemberInfo;
+import roomescape.service.dto.request.reservation.ReservationRequest;
+import roomescape.service.dto.request.reservation.ReservationSearchCond;
+import roomescape.service.dto.request.reservation.UserReservationRequest;
+import roomescape.service.dto.response.reservation.ReservationResponse;
+import roomescape.service.dto.response.reservation.UserReservationResponse;
 
 @RestController
 @RequiredArgsConstructor
