--- src/main/java/roomescape/web/dto/request/member/MemberInfo.java
@@ -1,4 +1,0 @@
-package roomescape.web.dto.request.member;
-
-public record MemberInfo(Long id) {
-}
