+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -20,8 +20,8 @@ import roomescape.domain.repository.MemberRepository;
 import roomescape.domain.repository.ReservationRepository;
 import roomescape.domain.repository.ReservationTimeRepository;
 import roomescape.domain.repository.ThemeRepository;
-import roomescape.web.dto.request.reservation.ReservationSearchCond;
-import roomescape.web.dto.response.reservation.ReservationResponse;
+import roomescape.service.dto.request.reservation.ReservationSearchCond;
+import roomescape.service.dto.response.reservation.ReservationResponse;
 
 @SpringBootTest
 class ReservationServiceTest {
