+++ src/test/java/roomescape/web/controller/AuthControllerTest.java
@@ -23,8 +23,8 @@ import roomescape.domain.Member;
 import roomescape.domain.Role;
 import roomescape.service.MemberService;
 import roomescape.service.security.JwtProvider;
-import roomescape.web.dto.request.member.LoginRequest;
-import roomescape.web.dto.response.member.MemberResponse;
+import roomescape.service.dto.request.member.LoginRequest;
+import roomescape.service.dto.response.member.MemberResponse;
 
 
 @WebMvcTest(controllers = AuthController.class)
